Update TeleCardEmulator.ino
I have abandoned the use of an ISR for sending card bits and have decided to use a simple flag process in the main loop.  These cards only run at about 15kHz in any event so it's not a big deal.  The next step is to clean up and do some maths to get the bits sent out correctly.